checking for id_rsa.pub

+++ docker/alpine/startup.sh
@@ -15,7 +15,16 @@ if [ ! -z "$DOCKER_COMPOSE" ]; then
   fi
 
   # Add pipline's id key to authorized_keys
-  cp /root/.ssh/id_rsa.pub /root/.ssh/authorized_keys
+  if [ -f /root/.ssh/id_rsa.pub ]; then
+    echo "Add pipline's 'id_rsa.pub' key to authorized_keys"
+    cp /root/.ssh/id_rsa.pub /root/.ssh/authorized_keys
+  else
+    echo "Generating 'id_rsa.pub' and adding it to authorized_keys"
+    ssh-keygen -A
+    yes | ssh-keygen -t rsa -N "" -f ~/.ssh/id_rsa
+    cp /root/.ssh/id_rsa.pub /root/.ssh/authorized_keys
+    mkdir -p /var/run/sshd
+  fi
 
   # Add piep/container to known_hosts by looping over evry pipe/containers
   for pipe in $(echo $PIPES | tr "," "\n")
